перешел на лонг в deathtime

+++ includes/philo_main.h
@@ -21,7 +21,7 @@
 
 typedef struct s_data {
 	int					philos_cnt;
-	int					death_time;
+	long				death_time;
 	int					eat_time;
 	int					sleep_time;
 	int					limit_to_eat;
